--Factor out common code in numpy array viewer and  pandas dataframe viewer --remove ununsed variables and lambdify
