+++ python/src/META-INF/python-core-common.xml
@@ -799,11 +799,7 @@
       <add-to-group group-id="Internal"/>
     </action>
 
-    <action id="PyDebugger.ViewArray" class="com.jetbrains.python.debugger.array.PyViewArrayAction" text="View as Array">
-      <add-to-group group-id="XDebugger.ValueGroup" anchor="after" relative-to-action="Debugger.Tree.AddToWatches"/>
-    </action>
-
-    <action id="PyDebugger.ViewDataFrame" class="com.jetbrains.python.debugger.dataframe.PyViewDataFrameAction" text="View as DataFrame">
+    <action id="PyDebugger.ViewArray" class="com.jetbrains.python.debugger.containerview.PyViewNumericContainerAction" text="View as Array">
       <add-to-group group-id="XDebugger.ValueGroup" anchor="after" relative-to-action="Debugger.Tree.AddToWatches"/>
     </action>
