--- python/src/com/jetbrains/python/debugger/array/ArrayTableForm.form
@@ -1,80 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="com.jetbrains.python.debugger.array.ArrayTableForm">
-  <grid id="27dc6" binding="myMainPanel" layout-manager="GridLayoutManager" row-count="2" column-count="5" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
-    <margin top="0" left="0" bottom="0" right="0"/>
-    <constraints>
-      <xy x="20" y="20" width="608" height="375"/>
-    </constraints>
-    <properties/>
-    <border type="none"/>
-    <children>
-      <scrollpane id="90774" class="com.intellij.ui.components.JBScrollPane" binding="myScrollPane" custom-create="true">
-        <constraints>
-          <grid row="0" column="0" row-span="1" col-span="5" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <horizontalScrollBarPolicy value="30"/>
-          <verticalScrollBarPolicy value="20"/>
-        </properties>
-        <border type="none"/>
-        <children>
-          <component id="c7f7c" class="com.intellij.ui.table.JBTable" binding="myTable" custom-create="true">
-            <constraints/>
-            <properties>
-              <autoResizeMode value="0"/>
-              <fillsViewportHeight value="true"/>
-            </properties>
-          </component>
-        </children>
-      </scrollpane>
-      <component id="b6d58" class="javax.swing.JCheckBox" binding="myColoredCheckbox">
-        <constraints>
-          <grid row="1" column="4" row-span="1" col-span="1" vsize-policy="0" hsize-policy="3" anchor="8" fill="0" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <enabled value="true"/>
-          <selected value="true"/>
-          <text value="Colored"/>
-        </properties>
-      </component>
-      <grid id="db2ae" binding="myFormatPanel" layout-manager="GridLayoutManager" row-count="1" column-count="2" same-size-horizontally="true" same-size-vertically="false" hgap="-1" vgap="-1">
-        <margin top="0" left="0" bottom="0" right="0"/>
-        <constraints>
-          <grid row="1" column="3" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties/>
-        <border type="none"/>
-        <children>
-          <component id="8f575" class="javax.swing.JLabel" binding="myFormatLabel">
-            <constraints>
-              <grid row="0" column="0" row-span="1" col-span="1" vsize-policy="0" hsize-policy="0" anchor="4" fill="0" indent="0" use-parent-layout="false"/>
-            </constraints>
-            <properties>
-              <horizontalAlignment value="4"/>
-              <horizontalTextPosition value="4"/>
-              <text value="Format:"/>
-            </properties>
-          </component>
-          <component id="821a3" class="com.intellij.ui.EditorTextField" binding="myFormatTextField" custom-create="true">
-            <constraints>
-              <grid row="0" column="1" row-span="1" col-span="1" vsize-policy="0" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false">
-                <minimum-size width="70" height="-1"/>
-              </grid>
-            </constraints>
-            <properties>
-              <toolTipText value="Format"/>
-            </properties>
-          </component>
-        </children>
-      </grid>
-      <component id="fd814" class="com.intellij.ui.EditorTextField" binding="mySliceTextField" custom-create="true">
-        <constraints>
-          <grid row="1" column="0" row-span="1" col-span="3" vsize-policy="0" hsize-policy="6" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
-        </constraints>
-        <properties>
-          <toolTipText value="Current slice."/>
-        </properties>
-      </component>
-    </children>
-  </grid>
-</form>
