+++ python/src/com/jetbrains/python/debugger/array/NumpyArrayTable.java
@@ -14,202 +14,83 @@
  * limitations under the License.
  */
 package com.jetbrains.python.debugger.array;
-
-import com.intellij.codeInsight.hint.HintManager;
-import com.intellij.openapi.application.ApplicationManager;
-import com.intellij.openapi.diagnostic.Logger;
-import com.intellij.openapi.editor.RangeMarker;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.util.Pair;
 import com.intellij.util.ui.UIUtil;
-import com.jetbrains.python.debugger.*;
+import com.jetbrains.python.debugger.ArrayChunk;
+import com.jetbrains.python.debugger.PyDebugValue;
+import com.jetbrains.python.debugger.containerview.NumericContainerViewTable;
+import com.jetbrains.python.debugger.containerview.ViewNumericContainerDialog;
 import org.jetbrains.annotations.NotNull;
 
-import javax.swing.*;
 import java.awt.*;
-import java.awt.event.*;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
+import java.awt.event.KeyAdapter;
+import java.awt.event.KeyListener;
+
 
 /**
  * @author amarch
  */
-public class NumpyArrayTable implements TableChunkDatasource {
-  private final PyDebugValue myValue;
-  private final PyViewArrayAction.ViewArrayDialog myDialog;
-  private final ArrayTableForm myComponent;
-  private final JTable myTable;
-  private Project myProject;
-  private PyDebuggerEvaluator myEvaluator;
-  private String myDtypeKind;
-  private ArrayTableCellRenderer myTableCellRenderer;
-  private AsyncArrayTableModel myPagingModel;
+public final class NumpyArrayTable extends NumericContainerViewTable {
 
-  private final static int COLUMNS_IN_DEFAULT_SLICE = 40;
-  private final static int ROWS_IN_DEFAULT_SLICE = 40;
 
-  private final static int COLUMNS_IN_DEFAULT_VIEW = 1000;
-  private final static int ROWS_IN_DEFAULT_VIEW = 1000;
-
-  private static final Pattern PY_COMPLEX_NUMBER = Pattern.compile("([+-]?[.\\d^j]*)([+-]?[e.\\d]*j)?");
-
-  private final static int HUGE_ARRAY_SIZE = 1000 * 1000;
-  private final static String LOAD_SMALLER_SLICE = "Full slice too large and would slow down debugger, shrunk to smaller slice.";
-  private final static String DISABLE_COLOR_FOR_HUGE_ARRAY =
-    "Disable color because array too big and calculating min and max would slow down debugging.";
-
-  private static final Logger LOG = Logger.getInstance("#com.jetbrains.python.debugger.array.NumpyArrayValueProvider");
-
+  private ArrayTableCellRenderer myArrayTableCellRenderer;
   public NumpyArrayTable(@NotNull Project project,
-                         @NotNull PyViewArrayAction.ViewArrayDialog dialog, @NotNull PyDebugValue value) {
-    myValue = value;
-    myDialog = dialog;
-    myComponent = new ArrayTableForm(project, new KeyAdapter() {
-      @Override
-      public void keyPressed(KeyEvent e) {
-        if (e.getKeyCode() == KeyEvent.VK_ENTER) {
-          doReslice();
-        }
-      }
-    }, new KeyAdapter() {
-      @Override
-      public void keyPressed(KeyEvent e) {
-        if (e.getKeyCode() == KeyEvent.VK_ENTER) {
-          doApplyFormat();
-        }
-      }
-    });
-    myTable = myComponent.getTable();
-    myProject = project;
-    myEvaluator = new PyDebuggerEvaluator(project, getDebugValue().getFrameAccessor());
+                         @NotNull ViewNumericContainerDialog dialog, @NotNull PyDebugValue value) {
+    super(project, dialog, value);
   }
 
-  public ArrayTableForm getComponent() {
-    return myComponent;
-  }
-
-  private void initComponent() {
-    myComponent.getColoredCheckbox().setEnabled(false);
-    myComponent.getColoredCheckbox().addItemListener(new ItemListener() {
-      @Override
-      public void itemStateChanged(ItemEvent e) {
-        if (e.getSource() == myComponent.getColoredCheckbox()) {
-
-          if (myTable.getRowCount() > 0 &&
-              myTable.getColumnCount() > 0 &&
-              myTable.getCellRenderer(0, 0) instanceof ArrayTableCellRenderer) {
-            ArrayTableCellRenderer renderer = (ArrayTableCellRenderer)myTable.getCellRenderer(0, 0);
-            if (myComponent.getColoredCheckbox().isSelected()) {
-              renderer.setColored(true);
-            }
-            else {
-              renderer.setColored(false);
-            }
-          }
-          myComponent.getScrollPane().repaint();
-        }
-      }
-    });
-
-    //make value name read-only
-    myComponent.getSliceTextField().addFocusListener(new FocusListener() {
-      @Override
-      public void focusGained(FocusEvent e) {
-        myComponent.getSliceTextField().getDocument().createGuardedBlock(0, getNodeFullName().length());
-      }
-
-      @Override
-      public void focusLost(FocusEvent e) {
-        RangeMarker block = myComponent.getSliceTextField().getDocument().getRangeGuard(0, getNodeFullName().length());
-        if (block != null) {
-          myComponent.getSliceTextField().getDocument().removeGuardedBlock(block);
-        }
-      }
-    });
+  @Override
+  protected final String getTitlePresentation(String slice) {
+    return "Array View: " + slice;
   }
 
   public void disableColor() {
-    myTableCellRenderer.setMin(Double.MAX_VALUE);
-    myTableCellRenderer.setMax(Double.MIN_VALUE);
+    myArrayTableCellRenderer.setMin(Double.MAX_VALUE);
+    myArrayTableCellRenderer.setMax(Double.MIN_VALUE);
     myTableCellRenderer.setColored(false);
     UIUtil.invokeLaterIfNeeded(() -> {
       myComponent.getColoredCheckbox().setSelected(false);
       myComponent.getColoredCheckbox().setEnabled(false);
       if (myTable.getColumnCount() > 0) {
-        myTable.setDefaultRenderer(myTable.getColumnClass(0), myTableCellRenderer);
+        myTable.setDefaultRenderer(myTable.getColumnClass(0), myArrayTableCellRenderer);
       }
     });
   }
 
-  public PyDebugValue getDebugValue() {
-    return myValue;
-  }
 
-  public PyFrameAccessor getEvaluator() {
-    return myValue.getFrameAccessor();
-  }
+  @Override
+  protected final void initUi(@NotNull final ArrayChunk chunk, final boolean inPlace) {
+    myPagingModel = new AsyncArrayTableModel(Math.min(chunk.getRows(), ROWS_IN_DEFAULT_VIEW),
+                                             Math.min(chunk.getColumns(), COLUMNS_IN_DEFAULT_VIEW), this);
+    myPagingModel.addToCache(chunk);
+    myDtypeKind = chunk.getType();
 
-  public void init() {
-    init(getDebugValue().getEvaluationExpression(), false);
-  }
-
-  public void init(final String slice, final boolean inPlace) {
-    initComponent();
-
-    ApplicationManager.getApplication().executeOnPooledThread(() -> {
-      final PyDebugValue value = getDebugValue();
-      PyDebugValue parent = value.getParent();
-      final PyDebugValue slicedValue =
-        new PyDebugValue(slice, value.getType(), null, value.getValue(), value.isContainer(), value.isReturnedVal(),
-                         value.isErrorOnEval(), parent, value.getFrameAccessor());
-
-      final String format = getFormat().isEmpty() ? "%" : getFormat();
-
-      try {
-        initUi(value.getFrameAccessor()
-                 .getArrayItems(slicedValue, 0, 0, -1, -1, format), inPlace);
+    UIUtil.invokeLaterIfNeeded(() -> {
+      myTable.setModel(myPagingModel);
+      myComponent.getSliceTextField().setText(chunk.getSlicePresentation());
+      myComponent.getFormatTextField().setText(chunk.getFormat());
+      myDialog.setTitle(getTitlePresentation(chunk.getSlicePresentation()));
+      myArrayTableCellRenderer = new ArrayTableCellRenderer(Double.MIN_VALUE, Double.MIN_VALUE, chunk.getType());
+      myTableCellRenderer = myArrayTableCellRenderer;
+      fillColorRange(chunk.getMin(), chunk.getMax());
+      if (!isNumeric()) {
+        disableColor();
       }
-      catch (PyDebuggerException e) {
-        showError(e.getMessage());
+      else {
+        myComponent.getColoredCheckbox().setEnabled(true);
+      }
+
+      if (!inPlace) {
+        myComponent.getScrollPane().getViewport().setViewPosition(new Point(0, 0));
+        JBTableWithRowHeaders.RowHeaderTable rowTable = ((JBTableWithRowHeaders)myTable).getRowHeaderTable();
+        rowTable.setRowShift(0);
+      }
+      ((AsyncArrayTableModel)myTable.getModel()).fireTableDataChanged();
+      ((AsyncArrayTableModel)myTable.getModel()).fireTableCellUpdated(0, 0);
+      if (myTable.getColumnCount() > 0) {
+        myTable.setDefaultRenderer(myTable.getColumnClass(0), myArrayTableCellRenderer);
       }
     });
-  }
-
-  private void initUi(@NotNull final ArrayChunk chunk, final boolean inPlace) {
-      myPagingModel = new AsyncArrayTableModel(Math.min(chunk.getRows(), ROWS_IN_DEFAULT_VIEW),
-                                               Math.min(chunk.getColumns(), COLUMNS_IN_DEFAULT_VIEW), this);
-      myPagingModel. addToCache(chunk);
-      myDtypeKind = chunk.getType();
-
-      UIUtil.invokeLaterIfNeeded(() -> {
-        myTable.setModel(myPagingModel);
-        myComponent.getSliceTextField().setText(chunk.getSlicePresentation());
-        myComponent.getFormatTextField().setText(chunk.getFormat());
-        myDialog.setTitle(getTitlePresentation(chunk.getSlicePresentation()));
-        myTableCellRenderer = new ArrayTableCellRenderer(Double.MIN_VALUE, Double.MIN_VALUE, chunk.getType());
-        fillColorRange(chunk.getMin(), chunk.getMax());
-        if (!isNumeric()) {
-          disableColor();
-        }
-        else {
-          myComponent.getColoredCheckbox().setEnabled(true);
-        }
-
-        if (!inPlace) {
-          myComponent.getScrollPane().getViewport().setViewPosition(new Point(0, 0));
-          JBTableWithRowHeaders.RowHeaderTable rowTable = ((JBTableWithRowHeaders)myTable).getRowHeaderTable();
-          rowTable.setRowShift(0);
-        }
-        ((AsyncArrayTableModel)myTable.getModel()).fireTableDataChanged();
-        ((AsyncArrayTableModel)myTable.getModel()).fireTableCellUpdated(0, 0);
-        if (myTable.getColumnCount() > 0) {
-          myTable.setDefaultRenderer(myTable.getColumnClass(0), myTableCellRenderer);
-        }
-      });
-  }
-
-  private static String getTitlePresentation(String slice) {
-    return "Array View: " + slice;
   }
 
   private void fillColorRange(String minValue, String maxValue) {
@@ -218,8 +99,8 @@ public class NumpyArrayTable implements TableChunkDatasource {
     if ("c".equals(myDtypeKind)) {
       min = 0;
       max = 1;
-      myTableCellRenderer.setComplexMin(minValue);
-      myTableCellRenderer.setComplexMax(maxValue);
+      myArrayTableCellRenderer.setComplexMin(minValue);
+      myArrayTableCellRenderer.setComplexMax(maxValue);
     }
     else if ("b".equals(myDtypeKind)) {
       min = minValue.equals("True") ? 1 : 0;
@@ -230,23 +111,11 @@ public class NumpyArrayTable implements TableChunkDatasource {
       max = Double.parseDouble(maxValue);
     }
 
-    myTableCellRenderer.setMin(min);
-    myTableCellRenderer.setMax(max);
-  }
-
-  public String getSliceText() {
-    return myComponent.getSliceTextField().getText();
+    myArrayTableCellRenderer.setMin(min);
+    myArrayTableCellRenderer.setMax(max);
   }
 
   @Override
-  public ArrayChunk getChunk(int rowOffset, int colOffset, int rows, int cols) throws PyDebuggerException {
-    final PyDebugValue slicedValue =
-      new PyDebugValue(getSliceText(), myValue.getType(), myValue.getTypeQualifier(), myValue.getValue(), myValue.isContainer(), myValue.isErrorOnEval(),
-                       myValue.getParent(), myValue.getFrameAccessor());
-
-    return myValue.getFrameAccessor().getArrayItems(slicedValue, rowOffset, colOffset, rows, cols, getFormat());
-  }
-
   public boolean isNumeric() {
     if (myDtypeKind != null) {
       return "biufc".contains(myDtypeKind.substring(0, 1));
@@ -254,7 +123,8 @@ public class NumpyArrayTable implements TableChunkDatasource {
     return false;
   }
 
-  private void initTableModel(final boolean inPlace) {
+  @Override
+  protected final void initTableModel(final boolean inPlace) {
     myPagingModel = new AsyncArrayTableModel(myPagingModel.getRowCount(), myPagingModel.getColumnCount(), this);
 
     UIUtil.invokeLaterIfNeeded(() -> {
@@ -267,119 +137,14 @@ public class NumpyArrayTable implements TableChunkDatasource {
       ((AsyncArrayTableModel)myTable.getModel()).fireTableDataChanged();
       ((AsyncArrayTableModel)myTable.getModel()).fireTableCellUpdated(0, 0);
       if (myTable.getColumnCount() > 0) {
-        myTable.setDefaultRenderer(myTable.getColumnClass(0), myTableCellRenderer);
+        myTable.setDefaultRenderer(myTable.getColumnClass(0), myArrayTableCellRenderer);
       }
     });
   }
 
   @Override
-  public String correctStringValue(@NotNull Object value) {
-    if (value instanceof String) {
-      String corrected = (String)value;
-      if ( isNumeric()) {
-        if (corrected.startsWith("\'") || corrected.startsWith("\"")) {
-          corrected = corrected.substring(1, corrected.length() - 1);
-        }
-      }
-      return corrected;
-    }
-    else if (value instanceof Integer) {
-      return Integer.toString((Integer)value);
-    }
-    return value.toString();
-  }
-
-  public void setDtypeKind(String dtype) {
-    this.myDtypeKind = dtype;
-  }
-
-  @Override
-  public void showError(String message) {
-    myDialog.setError(message);
-  }
-
-  public void showInfoHint(final String message) {
-    UIUtil.invokeLaterIfNeeded(() -> {
-      if (myComponent.getSliceTextField().getEditor() != null) {
-        HintManager.getInstance().showInformationHint(myComponent.getSliceTextField().getEditor(), message);
-      }
-    });
-  }
-
-  private void doReslice() {
-    clearErrorMessage();
-    init(getSliceText(), false);
-  }
-
-  private void clearErrorMessage() {
-    showError(null);
-  }
-
-  private void doApplyFormat() {
-    reset();
-  }
-
-  private void reset() {
-    clearErrorMessage();
-    initTableModel(true);
-  }
-
-  /**
-   * @return double presentation from [0:1] range
-   */
-  public static double getRangedValue(String value, String type, double min, double max, String complexMax, String complexMin) {
-    if ("iuf".contains(type)) {
-      return (Double.parseDouble(value) - min) / (max - min);
-    }
-    else if ("b".equals(type)) {
-      return value.equals("True") ? 1 : 0;
-    }
-    else if ("c".equals(type)) {
-      return getComplexRangedValue(value, complexMax, complexMin);
-    }
-    return 0;
-  }
-
-  /**
-   * type complex128 in numpy is compared by next rule:
-   * A + Bj > C +Dj if A > C or A == C and B > D
-   */
-  private static double getComplexRangedValue(String value, String complexMax, String complexMin) {
-    Pair<Double, Double> med = parsePyComplex(value);
-    Pair<Double, Double> max = parsePyComplex(complexMax);
-    Pair<Double, Double> min = parsePyComplex(complexMin);
-    double range = (med.first - min.first) / (max.first - min.first);
-    if (max.first.equals(min.first)) {
-      range = (med.second - min.second) / (max.second - min.second);
-    }
-    return range;
-  }
-
-  private static Pair<Double, Double> parsePyComplex(@NotNull String pyComplexValue) {
-    if (pyComplexValue.startsWith("(") && pyComplexValue.endsWith(")")) {
-      pyComplexValue = pyComplexValue.substring(1, pyComplexValue.length() - 1);
-    }
-    Matcher matcher = PY_COMPLEX_NUMBER.matcher(pyComplexValue);
-    if (matcher.matches()) {
-      String real = matcher.group(1);
-      String imag = matcher.group(2);
-      if (real.contains("j") && imag == null) {
-        return new Pair(new Double(0.0), Double.parseDouble(real.substring(0, real.length() - 1)));
-      }
-      else {
-        return new Pair(Double.parseDouble(real), Double.parseDouble(imag.substring(0, imag.length() - 1)));
-      }
-    }
-    else {
-      throw new IllegalArgumentException("Not a valid python complex value: " + pyComplexValue);
-    }
-  }
-
-  public String getNodeFullName() {
-    return getDebugValue().getEvaluationExpression();
-  }
-
-  public String getFormat() {
-    return myComponent.getFormatTextField().getText();
+  @NotNull
+  protected ArrayTableForm createForm(@NotNull Project project, KeyListener resliceCallback, KeyAdapter formatCallback) {
+    return new ArrayTableForm(project,resliceCallback, formatCallback);
   }
 }
